+++ aiobotocore/session.py
@@ -10,6 +10,7 @@ from botocore.session import (
     logger,
 )
 from botocore.session import Session as _SyncSession
+from botocore.useragent import register_feature_id
 
 from . import __version__, retryhandler
 from .client import AioBaseClient, AioClientCreator
@@ -191,10 +192,13 @@ class AioSession(_SyncSession):
                 aws_session_token, aws_account_id
             ):
                 logger.debug(
-                    f"Ignoring the following credential-related values which were set without "
-                    f"an access key id and secret key on the session or client: {ignored_credentials}"
+                    "Ignoring the following credential-related values which were set without "
+                    "an access key id and secret key on the session or client: %s",
+                    ignored_credentials,
                 )
             credentials = await self.get_credentials()
+        if getattr(credentials, 'method', None) == 'explicit':
+            register_feature_id('CREDENTIALS_CODE')
         auth_token = self.get_auth_token()
         endpoint_resolver = self._get_internal_component('endpoint_resolver')
         exceptions_factory = self._get_internal_component('exceptions_factory')