+++ tests/botocore_tests/functional/test_useragent.py
@@ -1,4 +1,4 @@
-import asyncio
+import anyio
 
 from aiobotocore.session import AioSession
 
@@ -93,6 +93,8 @@ async def test_registered_feature_ids_dont_bleed_across_threads():
             aws_access_key_id='xxx',
         ) as s3_client,
     ):
+        waiter_features = []
+        paginator_features = []
 
         async def wait():
             with ClientHTTPStubber(s3_client) as stub_client:
@@ -101,7 +103,7 @@ async def test_registered_feature_ids_dont_bleed_across_threads():
                 # The `wait()` method registers `'WAITER': 'B'`
                 await waiter.wait(Bucket='mybucket')
             ua_string = get_captured_ua_strings(stub_client)[0]
-            return parse_registered_feature_ids(ua_string)
+            waiter_features.extend(parse_registered_feature_ids(ua_string))
 
         async def paginate():
             with ClientHTTPStubber(s3_client) as stub_client:
@@ -111,11 +113,11 @@ async def test_registered_feature_ids_dont_bleed_across_threads():
                 async for _ in paginator.paginate():
                     pass
             ua_string = get_captured_ua_strings(stub_client)[0]
-            return parse_registered_feature_ids(ua_string)
+            paginator_features.extend(parse_registered_feature_ids(ua_string))
 
-        waiter_features, paginator_features = await asyncio.gather(
-            wait(), paginate()
-        )
+        async with anyio.create_task_group() as tg:
+            tg.start_soon(wait)
+            tg.start_soon(paginate)
 
         assert 'B' in waiter_features
         assert 'C' not in waiter_features