+++ tests/test_patches.py
@@ -168,7 +168,7 @@ def test_protocol_parsers():
         (
             ClientArgsCreator.get_client_args,
             {
-                'd651b5c3d81738b7f2072b984eef8b5d4cfacd22',
+                '3dfe4ed858d06ca35b5033c0f5dc9531e12090c2',
             },
         ),
         (
@@ -407,13 +407,13 @@ def test_protocol_parsers():
         (
             SSOCredentialFetcher._get_credentials,
             {
-                '13ac3b73e0745dfeaa934a8873179ca6c22a164f',
+                '6fdbe9d6045e98be5125fade72fb1711936666b0',
             },
         ),
         (
             SSOProvider.load,
             {
-                '67aba81dd1def437f2035f5e20b0720b328d970a',
+                'ca09b93ac1aea3df424652536d70f0fc52d86937',
             },
         ),
         (
@@ -437,7 +437,7 @@ def test_protocol_parsers():
         (
             AssumeRoleCredentialFetcher._get_credentials,
             {
-                '51e265bebde5b193d79a680706f2c7efc7b56622',
+                '7216e35a401c7bd95f8d95d872f55590ade8aaa7',
             },
         ),
         (
@@ -456,7 +456,7 @@ def test_protocol_parsers():
         (
             AssumeRoleWithWebIdentityCredentialFetcher._get_credentials,
             {
-                '23a7efbaf2022dcf713399649936cd9c2116f6c5',
+                '726f644df3f8e71125c8890193f2fa25bd22e331',
             },
         ),
         (
@@ -469,7 +469,7 @@ def test_protocol_parsers():
         (
             EnvProvider.load,
             {
-                '9c6bb8f48ea4cb3dfbe7df925a269a077205d62b',
+                'da0cfb81a7a27ac1159c9b87920908944d1849f2',
             },
         ),
         (
@@ -493,13 +493,13 @@ def test_protocol_parsers():
         (
             ContainerProvider._create_fetcher,
             {
-                'f61e3e60fda55de7dbcf4a66865981f7cb0804b8',
+                'f70e2b62a83099aa9863ed31fae7416b8a4c8df5',
             },
         ),
         (
             InstanceMetadataProvider.load,
             {
-                '15becfc0373ccfbc1bb200bd6a34731e61561d06',
+                '8b32da46500dd7b7177e1e6f781a7ce962dd3b66',
             },
         ),
         (
@@ -535,13 +535,13 @@ def test_protocol_parsers():
         (
             ConfigProvider.load,
             {
-                'dab41451970911f4b578015a3083ff550c6b7beb',
+                'af1d9e127b3d40376570cb48e3cbc8d66dff2041',
             },
         ),
         (
             SharedCredentialProvider.load,
             {
-                '833ec38b8361165e1bb285740e99c7ff7af161a5',
+                'fddbf6610526a41358187a529dbd2c2b82e23224',
             },
         ),
         (
@@ -553,7 +553,7 @@ def test_protocol_parsers():
         (
             ProcessProvider.load,
             {
-                '6691e82aeeccc4e07c30ee484ee31ebba5b6e2fb',
+                'dd8c5de56d32dd05f4e664faac84979ea84086cc',
             },
         ),
         (
@@ -577,7 +577,7 @@ def test_protocol_parsers():
         (
             AssumeRoleWithWebIdentityProvider._assume_role_with_web_identity,
             {
-                '32c9d720ab5f12054583758b5cd5d287f652ccd3',
+                '90c6a54d339f439988856f439dbaea9f371202f7',
             },
         ),
         (
@@ -589,19 +589,19 @@ def test_protocol_parsers():
         (
             AssumeRoleProvider._load_creds_via_assume_role,
             {
-                '85116d63561c9a8bfdfffdbf837b8a7e61b47ea3',
+                'bc9c38d6320b5f9712570ec8570f467e5288340d',
             },
         ),
         (
             AssumeRoleProvider._resolve_source_credentials,
             {
-                '105c0c011e23d76a3b8bd3d9b91b6d945c8307a1',
+                '1c2018d648dc72ab165cd1ffe1adc6b65856aa6b',
             },
         ),
         (
             AssumeRoleProvider._resolve_credentials_from_profile,
             {
-                '3e4e6371dabe4deb3445495aeb702eed8a3fba63',
+                'a5a2808e315b45957cb881b31b73624eba44ade6',
             },
         ),
         (
@@ -613,8 +613,7 @@ def test_protocol_parsers():
         (
             AssumeRoleProvider._resolve_credentials_from_source,
             {
-                'de41138b36bfc74d7f8a21f6002b55279d3de017',
-                '2f083f5c046de432194a2925fafc9478a5977586',
+                'b3144366432c11cee4bca286bd5ff10c3449b97b',
             },
         ),
         (
@@ -632,7 +631,7 @@ def test_protocol_parsers():
         (
             BotoProvider.load,
             {
-                'e84ebfe3d6698dc4683f0f6699d4a9907c87bebb',
+                'c85411fa3d3ff75b3edb98f38c36426c0ffed3bb',
             },
         ),
         (
@@ -907,8 +906,7 @@ def test_protocol_parsers():
         (
             EndpointRulesetResolver.construct_endpoint,
             {
-                'ccbed61e316a0e92e1d0f67c554ee15efa4ee6b8',
-                'ab22bb1ec171713e548567fbe84dd88a3d5f4b76',
+                '05e4f37e807b57bf9ff37dbe870308b684c62c02',
             },
         ),
         (
@@ -928,7 +926,7 @@ def test_protocol_parsers():
         (
             StreamingBody,
             {
-                'e6f0cb3b61c8b0a7c6961e77949e27c520b30a5c',
+                'e358f72191c4c1cb37d8fe90871342abf79afde2',
             },
         ),
         (
@@ -1065,8 +1063,7 @@ def test_protocol_parsers():
         (
             S3PostPresigner.generate_presigned_post,
             {
-                '269efc9af054a2fd2728d5b0a27db82c48053d7f',
-                '48418dc6c9b04fdc8689c7cb5b6eb987321a84e3',
+                '01c61dc0f33392a19def738afc634f5def401e32',
             },
         ),
         (
@@ -1158,13 +1155,13 @@ def test_protocol_parsers():
         (
             SSOTokenProvider._refresh_access_token,
             {
-                'cb179d1f262e41cc03a7c218e624e8c7fbeeaf19',
+                '6263c009e6d86011ca1226d61ea95163bbfca258',
             },
         ),
         (
             SSOTokenProvider._refresher,
             {
-                '824d41775dbb8a05184f6e9c7b2ea7202b72f2a9',
+                'd89b3446344826e5b740840c12fd9279810c45d2',
             },
         ),
         (
@@ -1285,7 +1282,7 @@ def test_protocol_parsers():
         (
             S3RegionRedirectorv2.redirect_from_error,
             {
-                '8e3003ec881c7eab0945fe4b6e021ca488fbcd78',
+                '2a715115e94bddcea4cce936bf7c7013f1f6ecdf',
             },
         ),
         (
@@ -1297,8 +1294,7 @@ def test_protocol_parsers():
         (
             S3RegionRedirector.redirect_from_error,
             {
-                '3863b2c6472513b7896bfccc9dfd2567c472f441',
-                'e1d93a4a85dfbfa810b9249da0b22ce14744b99d',
+                '5db904d0311db5c875aed0f6a78278f173ca40a7',
             },
         ),
         (
@@ -1416,7 +1412,7 @@ def test_protocol_parsers():
         (
             URLLib3Session,
             {
-                '1c418944abceb3a3d76c2c22348b4a39280d27ef',
+                'c65ca250524de59d89a32f213f38c7403666bfef',
             },
         ),
         (
@@ -1428,8 +1424,7 @@ def test_protocol_parsers():
         (
             EndpointDiscoveryManager.describe_endpoint,
             {
-                'b2f1b29177cf30f299e61b85ddec09eaa070e54e',
-                'cbd237b874daef01cf7be82fef30516557ba17f9',
+                '2d7c40eec571a14e6e3968b710c23677ae3685e6',
             },
         ),
         (
@@ -1524,7 +1519,7 @@ def test_protocol_parsers():
         (
             handle_checksum_body,
             {
-                '040cb48d8ebfb5ca195d41deb55b38d1fcb489f8',
+                'f019114f7fc3a4e200157b9689ed8a1cc9d72825',
             },
         ),
         (
@@ -1644,6 +1639,7 @@ def test_protocol_parsers():
             {
                 'bccf23c3733cc656b909f5130cba80dbc9540b05',
                 '7c01f505134b5ea3f4886e2288ea7f389577efd5',
+                '0ff1c068779d3e8a84c4da0655cfdf5861fe1b2c',
             },
         ),
     ],